+++ packages/leancode_contracts/CHANGELOG.md
@@ -4,6 +4,7 @@
 - **BREAKING**: Removed `TimeOnly.source`. Use `DateOnly.toDuration()` instead
 - **BREAKING**: `TimeOnly` constructor now accepts individual components instead of a `Duration` object. Use `TimeOnly.fromDuration` instead
 - Make `DateOnly` const-constructable
+- Allow `cqrs` v9
 
 # 0.5.0