+++ packages/leancode_contracts/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.6.0
+
+- **BREAKING**: Removed `DateOnly.source`. Use `DateOnly.toDateTimeLocal()` or `DateOnly.toDateTimeUtc()` instead
+- **BREAKING**: Removed `TimeOnly.source`. Use `DateOnly.toDuration()` instead
+- **BREAKING**: `TimeOnly` constructor now accepts individual components instead of a `Duration` object. Use `TimeOnly.fromDuration` instead
+- Allow `cqrs` v9
+
 # 0.5.0
 
 - **BREAKING**: Bump minimum dart SDK to 2.17.0